transcription/requirements.txt
fastapi
uvicorn[standard]
python-multipart
ffmpeg
transformers
openai-whisper
pandas
